requirements.txt
alembic==1.18.4
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.2
fonttools==4.61.1
fsspec==2026.3.0
idna==3.11
imbalanced-learn==0.14.1
imblearn==0.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kagglehub==1.0.0
kagglesdk==0.1.15
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
narwhals==2.17.0
nbclient==0.10.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
optuna==4.8.0
packaging==26.0
pandas==3.0.1
pandocfilters==1.5.1
parso==0.8.6
patsy==1.0.2
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
plotly==6.6.0
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==81.0.0
six==1.17.0
sklearn-compat==0.1.5
soupsieve==2.8.3
SQLAlchemy==2.0.48
stack-data==0.6.3
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
tinycss2==1.4.0
torch==2.11.0
torchinfo==1.8.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1
xgboost==3.2.0
